Combine schedulers
